Implement creating non-bare repositories - useful for hosting repositories where local access by scripts is needed.

+++ models/repo.go
@@ -914,6 +914,7 @@ func MigrateRepository(doer, u *User, opts MigrateRepoOptions) (*Repository, err
 		Description: opts.Description,
 		IsPrivate:   opts.IsPrivate,
 		IsMirror:    opts.IsMirror,
+		IsBare:      true,
 	})
 	if err != nil {
 		return nil, err
@@ -1031,7 +1032,7 @@ func cleanUpMigrateGitConfig(configPath string) error {
 }
 
 // createDelegateHooks creates all the hooks scripts for the repo
-func createDelegateHooks(repoPath string) (err error) {
+func createDelegateHooks(repoPath string, isBare bool) (err error) {
 	var (
 		hookNames = []string{"pre-receive", "update", "post-receive"}
 		hookTpls  = []string{
@@ -1046,7 +1047,12 @@ func createDelegateHooks(repoPath string) (err error) {
 		}
 	)
 
-	hookDir := filepath.Join(repoPath, "hooks")
+    var gitPath = repoPath
+	if !isBare {
+		gitPath = filepath.Join(repoPath, ".git")
+	}
+
+	hookDir := filepath.Join(gitPath, "hooks")
 
 	for i, hookName := range hookNames {
 		oldHookPath := filepath.Join(hookDir, hookName)
@@ -1072,11 +1078,11 @@ func createDelegateHooks(repoPath string) (err error) {
 // CleanUpMigrateInfo finishes migrating repository and/or wiki with things that don't need to be done for mirrors.
 func CleanUpMigrateInfo(repo *Repository) (*Repository, error) {
 	repoPath := repo.RepoPath()
-	if err := createDelegateHooks(repoPath); err != nil {
+	if err := createDelegateHooks(repoPath, repo.IsBare); err != nil {
 		return repo, fmt.Errorf("createDelegateHooks: %v", err)
 	}
 	if repo.HasWiki() {
-		if err := createDelegateHooks(repo.WikiPath()); err != nil {
+		if err := createDelegateHooks(repo.WikiPath(), true); err != nil {
 			return repo, fmt.Errorf("createDelegateHooks.(wiki): %v", err)
 		}
 	}
@@ -1094,7 +1100,7 @@ func CleanUpMigrateInfo(repo *Repository) (*Repository, error) {
 }
 
 // initRepoCommit temporarily changes with work directory.
-func initRepoCommit(tmpPath string, sig *git.Signature) (err error) {
+func initRepoCommit(tmpPath string, sig *git.Signature, isBare bool) (err error) {
 	var stderr string
 	if _, stderr, err = process.GetManager().ExecDir(-1,
 		tmpPath, fmt.Sprintf("initRepoCommit (git add): %s", tmpPath),
@@ -1109,10 +1115,12 @@ func initRepoCommit(tmpPath string, sig *git.Signature) (err error) {
 		return fmt.Errorf("git commit: %s", stderr)
 	}
 
-	if _, stderr, err = process.GetManager().ExecDir(-1,
-		tmpPath, fmt.Sprintf("initRepoCommit (git push): %s", tmpPath),
-		"git", "push", "origin", "master"); err != nil {
-		return fmt.Errorf("git push: %s", stderr)
+	if isBare {
+		if _, stderr, err = process.GetManager().ExecDir(-1,
+			tmpPath, fmt.Sprintf("initRepoCommit (git push): %s", tmpPath),
+			"git", "push", "origin", "master"); err != nil {
+			return fmt.Errorf("git push: %s", stderr)
+		}
 	}
 	return nil
 }
@@ -1126,6 +1134,7 @@ type CreateRepoOptions struct {
 	Readme      string
 	IsPrivate   bool
 	IsMirror    bool
+	IsBare      bool
 	AutoInit    bool
 }
 
@@ -1162,7 +1171,11 @@ func prepareRepoCommit(repo *Repository, tmpDir, repoPath string, opts CreateRep
 	if err != nil {
 		return fmt.Errorf("git clone: %v - %s", err, stderr)
 	}
+	return prepareRepositoryContent(repo, tmpDir, opts)
+}
 
+// Generate readme, licence and gitignore file
+func prepareRepositoryContent(repo *Repository, tmpDir string, opts CreateRepoOptions) error {
 	// README
 	data, err := getRepoInitFile("readme", opts.Readme)
 	if err != nil {
@@ -1224,10 +1237,10 @@ func initRepository(e Engine, repoPath string, u *User, repo *Repository, opts C
 		return fmt.Errorf("initRepository: path already exists: %s", repoPath)
 	}
 
-	// Init bare new repository.
-	if err = git.InitRepository(repoPath, true); err != nil {
+	// Init new repository.
+	if err = git.InitRepository(repoPath, opts.IsBare); err != nil {
 		return fmt.Errorf("InitRepository: %v", err)
-	} else if err = createDelegateHooks(repoPath); err != nil {
+	} else if err = createDelegateHooks(repoPath, opts.IsBare); err != nil {
 		return fmt.Errorf("createDelegateHooks: %v", err)
 	}
 
@@ -1236,18 +1249,25 @@ func initRepository(e Engine, repoPath string, u *User, repo *Repository, opts C
 	// Initialize repository according to user's choice.
 	if opts.AutoInit {
 
-		if err := os.MkdirAll(tmpDir, os.ModePerm); err != nil {
-			return fmt.Errorf("Failed to create dir %s: %v", tmpDir, err)
-		}
+		if opts.IsBare {
+			if err := os.MkdirAll(tmpDir, os.ModePerm); err != nil {
+				return fmt.Errorf("Failed to create dir %s: %v", tmpDir, err)
+			}
+	
+			defer os.RemoveAll(tmpDir)
 
-		defer os.RemoveAll(tmpDir)
-
-		if err = prepareRepoCommit(repo, tmpDir, repoPath, opts); err != nil {
-			return fmt.Errorf("prepareRepoCommit: %v", err)
+			if err = prepareRepoCommit(repo, tmpDir, repoPath, opts); err != nil {
+				return fmt.Errorf("prepareRepoCommit: %v", err)
+			}
+		} else {
+			if err = prepareRepositoryContent(repo, repoPath, opts); err != nil {
+				return fmt.Errorf("prepareRepository: %v", err)
+			}
+			tmpDir = repoPath
 		}
 
 		// Apply changes and commit.
-		if err = initRepoCommit(tmpDir, u.NewGitSig()); err != nil {
+		if err = initRepoCommit(tmpDir, u.NewGitSig(), opts.IsBare); err != nil {
 			return fmt.Errorf("initRepoCommit: %v", err)
 		}
 	}
@@ -1258,9 +1278,7 @@ func initRepository(e Engine, repoPath string, u *User, repo *Repository, opts C
 		return fmt.Errorf("getRepositoryByID: %v", err)
 	}
 
-	if !opts.AutoInit {
-		repo.IsBare = true
-	}
+	repo.IsBare = opts.IsBare
 
 	repo.DefaultBranch = "master"
 	if err = updateRepository(e, repo, false); err != nil {
@@ -1373,6 +1391,7 @@ func CreateRepository(doer, u *User, opts CreateRepoOptions) (_ *Repository, err
 		LowerName:   strings.ToLower(opts.Name),
 		Description: opts.Description,
 		IsPrivate:   opts.IsPrivate,
+		IsBare:      opts.IsBare,
 	}
 
 	sess := x.NewSession()
@@ -2129,11 +2148,12 @@ func ReinitMissingRepositories() error {
 func SyncRepositoryHooks() error {
 	return x.Where("id > 0").Iterate(new(Repository),
 		func(idx int, bean interface{}) error {
-			if err := createDelegateHooks(bean.(*Repository).RepoPath()); err != nil {
+			repo := bean.(*Repository)
+			if err := createDelegateHooks(repo.RepoPath(), repo.IsBare); err != nil {
 				return fmt.Errorf("SyncRepositoryHook: %v", err)
 			}
-			if bean.(*Repository).HasWiki() {
-				if err := createDelegateHooks(bean.(*Repository).WikiPath()); err != nil {
+			if repo.HasWiki() {
+				if err := createDelegateHooks(repo.WikiPath(), true); err != nil {
 					return fmt.Errorf("SyncRepositoryHook: %v", err)
 				}
 			}
@@ -2354,6 +2374,7 @@ func ForkRepository(doer, u *User, oldRepo *Repository, name, desc string) (_ *R
 		DefaultBranch: oldRepo.DefaultBranch,
 		IsPrivate:     oldRepo.IsPrivate,
 		IsFork:        true,
+		IsBare:        true,
 		ForkID:        oldRepo.ID,
 	}
 
@@ -2386,7 +2407,7 @@ func ForkRepository(doer, u *User, oldRepo *Repository, name, desc string) (_ *R
 		return nil, fmt.Errorf("git update-server-info: %v", stderr)
 	}
 
-	if err = createDelegateHooks(repoPath); err != nil {
+	if err = createDelegateHooks(repoPath, true); err != nil {
 		return nil, fmt.Errorf("createDelegateHooks: %v", err)
 	}
